move public user id to domain
