--- app/api/users/publicUser.ts
@@ -1,3 +1,0 @@
-import { ObjectId } from 'mongodb';
-
-export const PUBLIC_USER_ID = new ObjectId('698c35e7cf8880419d91fe4d');
